requirements.txt
azure.quantum
azure-quantum[cirq]
python-dotenv
matplotlib
azure-storage-blob
azure-identity
pytest
azure-cli